+++ Make_Draw_LaneDraw.h
@@ -12,9 +12,9 @@ namespace Make {
 		private:
 			const std::uint32_t laneColor;     //レーンの色
 			const std::uint32_t backColor;     //レーンの背景色
-			const std::uint8_t laneThickness;  //レーンの太さ
+			const std::uint16_t laneThickness;  //レーンの太さ
 			double laneWidth; //レーンの幅
-			std::uint8_t laneAmount;		   //レーンの数
+			std::uint16_t laneAmount;		   //レーンの数
 		public:
 			Make_Draw_LaneDraw();
 			void draw();
