+++ Make_File_MusicData.cpp
@@ -1,6 +1,6 @@
 #include "Make_File_MusicData.h"
 
-Make::File::Make_File_MusicData::Make_File_MusicData(const std::string& name,const std::string& artist,const std::uint8_t level,const double bp,const std::uint16_t bLength,const double tMinutes,const double bDlay) :
+Make::File::Make_File_MusicData::Make_File_MusicData(const std::string& name,const std::string& artist,const std::uint16_t level,const double bp,const std::uint16_t bLength,const double tMinutes,const double bDlay) :
 name(name),artist(artist),level(level),bpm(bp),barLength(bLength),totalMinutes(tMinutes),beginDelay(bDlay){}
 
 const std::string& Make::File::Make_File_MusicData::getName() noexcept {
@@ -11,7 +11,7 @@ const std::string& Make::File::Make_File_MusicData::getArtist() noexcept {
 	return artist;
 }
 
-const std::uint8_t& Make::File::Make_File_MusicData::getLevel() noexcept {
+const std::uint16_t& Make::File::Make_File_MusicData::getLevel() noexcept {
 	return level;
 }
