+++ Make_Note_LongNoteContainer.h
@@ -13,9 +13,9 @@ namespace Make {
 		private:
 			static double noteWidth; //描画するノーツの幅；
 			const std::uint16_t barID; //何小節目に属しているか
-			const std::uint8_t beatID;//その小節の何番目の線か
+			const std::uint16_t beatID;//その小節の何番目の線か
 			const double time;//曲の開始から何秒か
-			const std::uint8_t laneAmount; //レーンの数
+			const std::uint16_t laneAmount; //レーンの数
 			const double& r_y; //拍線の座標
 			std::uint32_t color;
 			std::vector<std::pair<bool,bool>> noteFlag;//firstはノーツがセットされているか、secondはそれが始点または終点か
@@ -26,18 +26,18 @@ namespace Make {
 			double notePointY;
 			std::vector<uint16_t> noteGroup;
 		public:
-			Make_Note_LongNoteContainer(const std::uint16_t barID,const std::uint8_t beatID,const double& y,std::uint8_t amountOfLane,const double time);
-			void setLongNoteFlag(const std::uint8_t laneID,const bool isFirstOrLast);//ノーツをセット既にセットされているなら撤去
+			Make_Note_LongNoteContainer(const std::uint16_t barID,const std::uint16_t beatID,const double& y,std::uint16_t amountOfLane,const double time);
+			void setLongNoteFlag(const std::uint16_t laneID,const bool isFirstOrLast);//ノーツをセット既にセットされているなら撤去
 			void drawLongNote();
-			void setNoteHeight(const std::uint8_t laneID,const double noteHeight,const bool isFirst);
-			void setNoteGroup(const std::uint8_t laneID,const std::uint16_t group);
+			void setNoteHeight(const std::uint16_t laneID,const double noteHeight,const bool isFirst);
+			void setNoteGroup(const std::uint16_t laneID,const std::uint16_t group);
 
-			const std::uint16_t& getNoteGroup(const std::uint8_t laneID);
+			const std::uint16_t& getNoteGroup(const std::uint16_t laneID);
 			const double& getY();
-			const std::pair<bool, bool> getLongNoteFlag(const std::uint8_t laneID);
+			const std::pair<bool, bool> getLongNoteFlag(const std::uint16_t laneID);
 			const double& getTime();
 			const std::uint16_t& getBarID();
-			const std::uint8_t& getBeatID();
+			const std::uint16_t& getBeatID();
 		};
 	}
 }
