+++ Make_Play_NormalNote.h
@@ -19,9 +19,9 @@ namespace Make {
 			const double time;//ノーツの時間
 			const double laneXRight;
 			const double laneXLeft;
-			const std::uint8_t noteType;
-			const std::uint8_t laneIndex;
-			const std::function<void(std::uint8_t , std::uint8_t)> nextNote;//判定を同じレーンの次のノーツに移す
+			const std::uint16_t noteType;
+			const std::uint16_t laneIndex;
+			const std::function<void(std::uint16_t , std::uint16_t)> nextNote;//判定を同じレーンの次のノーツに移す
 			const std::shared_ptr<Make_Play_Score>& p_score;//スコア表示
 			double y;
 			std::int32_t noteColor;
@@ -29,7 +29,7 @@ namespace Make {
 			bool done; //処理が終わったか
 			bool turn; //今自分の処理順か
 		public:
-			Make_Play_NormalNote(const double time, const std::uint8_t noteType, const std::uint8_t laneIndex, const double laneXRight,const double laneXLeft,const std::function<void(std::uint8_t, std::uint8_t)> nextNote, std::shared_ptr<Make_Play_Score>& p_score);
+			Make_Play_NormalNote(const double time, const std::uint16_t noteType, const std::uint16_t laneIndex, const double laneXRight,const double laneXLeft,const std::function<void(std::uint16_t, std::uint16_t)> nextNote, std::shared_ptr<Make_Play_Score>& p_score);
 			void check(double nowTime);
 			void setTurn(bool t);
 			void setDone(bool d);
